requirements.txt
python-telegram-bot
python-dotenv
ffmpeg
numpy
scipy
requests